Initial commit: A updated version after ROPTLIB 0.5
Compared to 0.5 version, the following is the updates:

1), update fixed-rank manifold. Add EucHvtoHv. Support sparse
representation of the Euclidean gradient.

2), Update the include directories of all the cpp files. Now, one does
not have to include all the subdirectories when creating a cpp project.
Only adding the root directory is enough.

3), Eucgrad->Riegrad->grad is usually more expensive than Eucgrad->grad
when intrinsic representation is used. In this version, Stiefel, Sphere,
OrthGroup, Oblique, Grassmann supports the efficient approach. One can
use Eucgrad->grad directly for these manifolds if one uncomments a few
codes in Problem:RieGrad().

4), Default of trust region max_delta is set to be 10000.

5), In linesearch algorithm, when intrinsic representation is used, then
the search direction is converted to extrinsic representation first.
This is usually faster than converting every time when applying
retraction.

6), Add Cayley retraction and Cayley vector transport in the Stiefel
manifold.

7), This version of ROPTLIB accepts a referenced iterate x_. The output
also includes an array which is dist(x_i, x_) for all i. The dist(x_i,
x_) is defined to be \|x_i - x_\|_F by default.

8), All functions std::cout are modified to be printf since Matlab does
not support std::cout.