Combine with FFTW and other updates
This version has been tested on all environments except Matlab in
Ubuntu. The methods of compiling ROPTLIB in different system therefore
need be updated. See the updated User Manual coming soon or send email
to huwst08@gmail.com

*） FFTW library is added. Users are able to invoke functions in FFTW
library.

*), 2 factor representation of a quotient manifold is added. This
manifold is equivalent to the fixed-rank complex manifold.

*), In retraction, the step size is also input as a parameter. This is
used in proximal mapping method.

*), In lowrank.cpp, fixed the bug that an error occurs when the fixed
rank is full rank.

*), Add manifold of the unit sphere on a tangent space of any given
manifold. This manifold is used to find the largest or smallest
eigvenvalue of the Hessian at any point.

*), Add Blind deconvolution problems. Three cost functions with
different domains are used. The first one is three factor representation
in embedded structure. The second one is two factor representation of
the total space. The third is the two factor representation of the
quotient manifold. The corresponding Cpp and Matlab driver and test
files are also added.

*), Phase retrieval problem is added. The corresponding Cpp and Matlab
driver and test files are also added.

*), Modify the default value for step size. BB2 is used in RSD. The
initial scaling of the Hessian in LRBFGS is also BB2 now.

*), In Matlab driver, users are able to find the largest and smallest
eigenvalues of the Hessian at the initial and final iterates in this
version.